+++ Makefile
@@ -1,0 +1,2 @@
+NesEmulator:
+	echo "Building NES Emulator Application"
