--- NesEmulator.sln
@@ -1,51 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.7.34003.232
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "NesEmulator", "NesEmulator\NesEmulator.vcxproj", "{1EA79101-0E0B-4C99-87D9-126ECA575B06}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Core", "Core\Core.vcxproj", "{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Gui", "Gui\Gui.vcxproj", "{9FF54174-0451-43C2-BBF3-847B24813F23}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1EA79101-0E0B-4C99-87D9-126ECA575B06}.Debug|x64.ActiveCfg = Debug|x64
-		{1EA79101-0E0B-4C99-87D9-126ECA575B06}.Debug|x64.Build.0 = Debug|x64
-		{1EA79101-0E0B-4C99-87D9-126ECA575B06}.Debug|x86.ActiveCfg = Debug|Win32
-		{1EA79101-0E0B-4C99-87D9-126ECA575B06}.Debug|x86.Build.0 = Debug|Win32
-		{1EA79101-0E0B-4C99-87D9-126ECA575B06}.Release|x64.ActiveCfg = Release|x64
-		{1EA79101-0E0B-4C99-87D9-126ECA575B06}.Release|x64.Build.0 = Release|x64
-		{1EA79101-0E0B-4C99-87D9-126ECA575B06}.Release|x86.ActiveCfg = Release|Win32
-		{1EA79101-0E0B-4C99-87D9-126ECA575B06}.Release|x86.Build.0 = Release|Win32
-		{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}.Debug|x64.ActiveCfg = Debug|x64
-		{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}.Debug|x64.Build.0 = Debug|x64
-		{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}.Debug|x86.ActiveCfg = Debug|Win32
-		{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}.Debug|x86.Build.0 = Debug|Win32
-		{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}.Release|x64.ActiveCfg = Release|x64
-		{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}.Release|x64.Build.0 = Release|x64
-		{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}.Release|x86.ActiveCfg = Release|Win32
-		{6716FF3D-768F-4CEA-A5E7-F1203DDC9692}.Release|x86.Build.0 = Release|Win32
-		{9FF54174-0451-43C2-BBF3-847B24813F23}.Debug|x64.ActiveCfg = Debug|x64
-		{9FF54174-0451-43C2-BBF3-847B24813F23}.Debug|x64.Build.0 = Debug|x64
-		{9FF54174-0451-43C2-BBF3-847B24813F23}.Debug|x86.ActiveCfg = Debug|Win32
-		{9FF54174-0451-43C2-BBF3-847B24813F23}.Debug|x86.Build.0 = Debug|Win32
-		{9FF54174-0451-43C2-BBF3-847B24813F23}.Release|x64.ActiveCfg = Release|x64
-		{9FF54174-0451-43C2-BBF3-847B24813F23}.Release|x64.Build.0 = Release|x64
-		{9FF54174-0451-43C2-BBF3-847B24813F23}.Release|x86.ActiveCfg = Release|Win32
-		{9FF54174-0451-43C2-BBF3-847B24813F23}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {61AEA412-BA0C-475B-932E-3117E95AE0EC}
-	EndGlobalSection
-EndGlobal
