--- NesEmulator/x64/Debug/NesEmulator.log
@@ -1,10 +1,0 @@
-﻿  Application.cpp
-  main.cpp
-  Generating Code...
-Application.obj : error LNK2019: unresolved external symbol "public: __cdecl Cartridge::Cartridge(class std::basic_string<char,struct std::char_traits<char>,class std::allocator<char> > const &)" (??0Cartridge@@QEAA@AEBV?$basic_string@DU?$char_traits@D@std@@V?$allocator@D@2@@std@@@Z) referenced in function "void __cdecl std::_Construct_in_place<class Cartridge,char const (&)[12]>(class Cartridge &,char const (&)[12])" (??$_Construct_in_place@VCartridge@@AEAY0M@$$CBD@std@@YAXAEAVCartridge@@AEAY0M@$$CBD@Z)
-Application.obj : error LNK2019: unresolved external symbol "public: __cdecl Cartridge::~Cartridge(void)" (??1Cartridge@@QEAA@XZ) referenced in function "public: void * __cdecl Cartridge::`scalar deleting destructor'(unsigned int)" (??_GCartridge@@QEAAPEAXI@Z)
-Application.obj : error LNK2019: unresolved external symbol "public: class olc::Sprite & __cdecl PPU::GetScreen(void)" (?GetScreen@PPU@@QEAAAEAVSprite@olc@@XZ) referenced in function "public: virtual bool __cdecl Application::OnUserUpdate(float)" (?OnUserUpdate@Application@@UEAA_NM@Z)
-Application.obj : error LNK2019: unresolved external symbol "public: void __cdecl Bus::insertCartridge(class std::shared_ptr<class Cartridge> const &)" (?insertCartridge@Bus@@QEAAXAEBV?$shared_ptr@VCartridge@@@std@@@Z) referenced in function "public: virtual bool __cdecl Application::OnUserCreate(void)" (?OnUserCreate@Application@@UEAA_NXZ)
-Application.obj : error LNK2019: unresolved external symbol "public: void __cdecl Bus::reset(void)" (?reset@Bus@@QEAAXXZ) referenced in function "public: virtual bool __cdecl Application::OnUserCreate(void)" (?OnUserCreate@Application@@UEAA_NXZ)
-Application.obj : error LNK2019: unresolved external symbol "public: void __cdecl Bus::clock(void)" (?clock@Bus@@QEAAXXZ) referenced in function "public: virtual bool __cdecl Application::OnUserUpdate(float)" (?OnUserUpdate@Application@@UEAA_NM@Z)
-D:\Programming\Projects2023\NesEmulator\x64\Debug\NesEmulator.exe : fatal error LNK1120: 6 unresolved externals
