--- NesEmulator/x64/Debug/NesEmulator.tlog/Cl.items.tlog
@@ -1,2 +1,0 @@
-D:\Programming\Projects2023\NesEmulator\NesEmulator\src\Application.cpp;D:\Programming\Projects2023\NesEmulator\NesEmulator\x64\Debug\Application.obj
-D:\Programming\Projects2023\NesEmulator\NesEmulator\src\main.cpp;D:\Programming\Projects2023\NesEmulator\NesEmulator\x64\Debug\main.obj
