--- NesEmulator/x64/Debug/NesEmulator.tlog/NesEmulator.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.37.32822:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|D:\Programming\Projects2023\NesEmulator\|
