--- NesEmulator/x64/Debug/NesEmulator.tlog/link.read.1.tlog
@@ -1,1 +1,0 @@
-ÿþ
